psppire-data-window: Set default filter in File|Save As to show all files.
The File|Save As dialog box showed "System Files" in two places: in the
filter dropdown and in the format dropdown.  It was surprising to users
that changing the filter dropdown had no effect on the format of the file
actually saved.  This commit makes it more obvious which dropdown needs to
be changed, by only showing "System Files" in the right place.

Reported by Harry Thijssen.
Bug #39648.