src/data/any-reader.c (any_reader_may_open): Return tri-state value.
Previously, any_reader_may_open returned bool.  This change makes it
return a tri-state variable. YES, NO or ERROR.  This is useful to
distinguish between a file which is not "readable" (ie not in a format
which PSPP knows how to read) and a file which cannot be opened due
to lack of permission or other lower level errors.