Change many %g format specifiers to %.*g with precision DBL_DIG + 1.
The high precision should ensure that in most cases the value is formatted
with enough digits to fully express the internal precision.  Without this
change, %g by itself only expresses 6 digits of precision, so that an
integer value like 99999999 is formatted as 1e8.  This fixes the problem.

This commit adds a test to MISSING VALUES because this was in particular
reported as a problem.

Thanks to Andre Müller <Andre.Mueller@gesis.org> for reporting this issue.