sys-file-reader: Don't skip "too many value labels" test anymore.
Previously the maximum number of value labels depended on the size of
size_t, so that it was impossible to have too many on 64-bit systems, but
a recent change made the maximum number of value labels depend on the
size of unsigned int instead, so now 32- and 64-bit systems are the same.
This commit reverts the code that skips the test.