Fixed bug importing ODS files.
There was a typo which caused the import of ODS files in the gui to
not properly function.  This change fixes the typo, but also re-arranges
the code such that similar errors are less likely to go unnoticed in the
future.  It also brings the code more in concord with the GNU Coding
Standards (which were written that way exactly to prevent these errors).

Fixes bug #41657