FREQUENCIES: Fixed crash showing histograms of limited range.
We had been using Freedman-Diaconis rule to select the bin width.
According to the literature this is better than Sturges rule.
However it cannot work when the interquartile range is zero, which
will happen for datasets with small range.

This change uses the Freedman-Diaconis rule and falls back to
Sturges when the IQR is zero.