render: Don't add overflows when header is contiguous with body.
render_page_select() needs to calculate overflows when it cuts a header off
from the body of a table.  Until now, it has considered headers and body to
always be cut off from each other, but this isn't true when the selected
body region is adjacent to one of the headers.  This commit takes that into
consideration, producing better results in one of the tests in the
testsuite.

This commit is much smaller when viewed with whitespace changes ignored.