render: Fix rendering corner cases when a cell is broken across pages.
The rendering code ability to break individual cells across multiple pages
has never received good test coverage because the criteria for doing that
are strict enough that the testsuite hardly ever does it.  This commit adds
a "--min-break=0" version to each test in the testsuite, which uncovers a
number of bugs in the cell breaking code.  It also fixes those bugs.