psppire-dialog-action-var-info: Only allow "Jump To" with one variable.
Until now, when at least one variable was selected, the user could click
on "Jump To", which would jump to the first selected variable.  This commit
changes behavior so that "Jump To" is only enabled when exactly one
variable is selected, which makes it a lot less confusing about which
variable will be jumped to.

Suggested by Bastián Díaz <diaz.bastian@ymail.com> .