zip-writer: Write size and CRC in local directory when possible.
Google Drive rejects ODT files as corrupted unless their size and CRC are
in local file headers, even though the zip file specification says that it
is OK for them to follow the file data.  Thus, this commit changes the
zip writer to satisfy Google unless the output file is not seekable.