output: Add footnote support.
The initials support for footnotes does not allow the client to specify
the markers and does not merge footnotes with the same text so that they
have the same markers.  It might be necessary to implement one or both of
those features.

The footnote feature doesn't have any users yet, outside of the tests.