cairo: Add .5pt gutter around table rules.
Commit d3fef25674baf4f4 (output: Add footnote support.) removed the 3-pt
gutter around table rules, instead putting a margin around cell contents.
This usually worked out fine, but for some tables in the GUI it caused the
rightmost rule in the table to disappear.  This is probably some kind of
off-by-one error in table size computation or clipping.  We may have to
eventually resolve that.  However, this commit seems like an adequate
temporary fix: instead of fixing the off-by-one error, add a thin gutter
around the rules, so that the actual content of the rule doesn't get
clipped off.