psppire-output-view: Fix segfault during printing.
psppire-output-view tried to make use of a special feature in the
cairo output driver that allowed the cairo_t to be omitted to avoid
actually rendering anything, for pagination purposes, but the cairo
code had changed and this feature had not been carefully maintained,
resulting in a segfault.  This commit stops trying to use that
feature, since it seems like it could easily get broken again in the
future and doesn't have much benefit anyhow.

Bug #43816.
Reported by John Darrington.